rename nav drawer simple demo files to match component path
