+++ src/app/navdrawer/nav-drawer-simple/nav-drawer-simple.component.ts
@@ -1,0 +1,9 @@
+import { Component } from "@angular/core";
+
+@Component({
+  selector: "app-nav-drawer-simple",
+  styleUrls: ["./nav-drawer-simple.component.css"],
+  templateUrl: "./nav-drawer-simple.component.html"
+})
+export class NavDrawerSimpleComponent  {
+}
