--- src/app/navdrawer/nav-drawer-simple/navdrawer-simple.component.ts
@@ -1,9 +1,0 @@
-import { Component } from "@angular/core";
-
-@Component({
-  selector: "app-navdrawer-simple",
-  styleUrls: ["./navdrawer-simple.component.css"],
-  templateUrl: "./navdrawer-simple.component.html"
-})
-export class NavDrawerSimpleComponent  {
-}
